requirements.txt
flask
pandas
mysql-connector-python
matplotlib